requirements.txt
einops==0.4.1
markdown2==2.4.6
matplotlib==3.5.2
numpy==1.23.1
opencv_python_headless==4.6.0.66
openpyxl==3.0.10
Pillow==9.3.0
pysodmetrics==1.3.1
PyYAML==6.0
scikit-image==0.19.2
tabulate==0.9.0
tensorboardX==2.5.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
